requirements.txt
google-cloud-securitycenter>=1.12.0
google-cloud-bigquery>=2.34.3
pandas>=1.4.2
reportlab>=3.6.8
pyyaml>=6.0
python-dateutil>=2.8.2
schedule>=1.2.0
pytest>=7.3.1
pytest-cov>=4.0.0
bandit>=1.7.5
safety>=2.3.5
requests>=2.31.0
google-cloud-storage>=2.9.0
google-auth>=2.22.0
python-dotenv>=1.0.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-gcp-trace>=1.4.0
opentelemetry-exporter-gcp-monitoring>=1.4.0
opentelemetry-instrumentation-requests>=0.39b0
prometheus-client>=0.16.0
google-cloud-monitoring>=2.14.1
google-cloud-trace>=1.11.1
datadog>=0.44.0
sentry-sdk>=1.28.1
elasticsearch>=8.8.0
fluent-logger>=0.9.7
